demo/requirements.txt
Django==1.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-bootstrap3==7.1.0
django-crispy-forms==1.6.1
django-user-accounts==2.0.0
MySQL-python==1.2.5
pinax-theme-bootstrap==7.11.0
pytz==2016.7
